Use uncompressed stylesheets by default
When we pushed a recent job live, everything went extremely white, and in a damn
hurry. The culprit was the absence of a minified/compressed stylesheet, one that
I'd forgotten to generate prior to the hasty deployment. Obviously this is going
to happen time-and-again with tight deadlines,  so the best thing to do would be
to disable the use of compressed assets by default (it should've been managed by
a WordPress setting, anyway...)